writte common_test file
