+++ builder/vsphere/iso/common_test.go
@@ -1,0 +1,17 @@
+package iso
+
+import (
+	"bytes"
+
+	"github.com/hashicorp/packer/helper/multistep"
+	"github.com/hashicorp/packer/packer"
+)
+
+func basicStateBag() *multistep.BasicStateBag {
+	state := new(multistep.BasicStateBag)
+	state.Put("ui", &packer.BasicUi{
+		Reader: new(bytes.Buffer),
+		Writer: new(bytes.Buffer),
+	})
+	return state
+}
